Extract tests of Data.List
